Refactored project into standard format. Closes #30.
