+++ src/main/java/com/dart/explore/entity/Amenity.java
@@ -1,5 +1,5 @@
 // Amenity.java
-package com.dart.entity;
+package com.dart.explore.entity;
 
 import javax.persistence.*;
