+++ src/main/java/com/dart/explore/main/ExploreApplication.java
@@ -1,4 +1,4 @@
-package com.dart.explore;
+package com.dart.explore.main;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
