+++ src/main/java/com/dart/explore/repository/AmenityRepository.java
@@ -1,6 +1,6 @@
-package com.dart.repository;
+package com.dart.explore.repository;
 
-import com.dart.entity.Amenity;
+import com.dart.explore.entity.Amenity;
 import org.springframework.data.repository.CrudRepository;
 
 public interface AmenityRepository extends CrudRepository<Amenity, Long> {
