+++ src/main/resources/application.properties
@@ -1,5 +1,6 @@
 
-spring.datasource.url=jdbc:postgresql://localhost:5432/DartApp
-spring.datasource.username=admin
+spring.datasource.url=jdbc:mysql://localhost:3306/dart_explore
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.username=root
 spring.datasource.password=password
 spring.jpa.hibernate.ddl-auto=update
